% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std.fin.my.R
\name{std.fin.my}
\alias{std.fin.my}
\title{std.fin.my = standardized financials multiple years. Returns professional-grade historical financial data. This data is standardized, cleansed and verified to ensure the highest quality data sourced directly from the XBRL financial statements.}
\usage{
std.fin.my(ticker = "AAPL", statement = "income_statement",
  period_type = "FY", period.quantity = 5, username = username,
  password = password)
}
\value{
Returns a list as table with the tag name and the information for a given year.
}
\description{
std.fin.my = standardized financials multiple years. Returns professional-grade historical financial data. This data is standardized, cleansed and verified to ensure the highest quality data sourced directly from the XBRL financial statements.
}
\examples{
data_avarage("AAPL", "cash_flow_statement", "FY", 3) #returns a the income statement for Apple Inc for the last 5 years


}
